+++ client/src/main/js/component/local/DetalhesLocalEdicao.js
@@ -4,6 +4,7 @@
     angular.module('localModulo').component('detalhesLocalEdicao', {
         templateUrl: 'view/component/local/detalhes-local-edicao.html',
         bindings: {
+            formulario: '=',
             local: '='
         },
         controller: [function () {