+++ client/src/main/js/component/local/LocalEdicao.js
@@ -4,6 +4,7 @@
     angular.module('localModulo').component('localEdicao', {
         templateUrl: 'view/component/local/local-edicao.html',
         bindings: {
+            formulario: '=',
             local: '='
         },
         controller: [function () {