+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Funcionalidade: Adicionada validação de permissão ao criar e alterar reservas no servidor.
 * Correção: Corrigidos testes no servidor com mudança de Secret do Auth0.
 * Funcionalidade: Adicionado modal com informações sobre os desenvolvedores e links para github.
+* Correção: Adicionada validação para intervalo em funcionamento de local.
 
 # 0.0.2
 * Início versionamento.